backend/requirements.txt
flask
flask-cors
youtube-transcript-api
groq
python-dotenv
gunicorn
requests